框架：转换long主键 wojilu.apps.download

+++ wojilu.Apps.Download/Controller/Admin/CategoryController.cs
@@ -29,7 +29,7 @@ namespace wojilu.Web.Controller.Download.Admin {
             bindList( "list", "data", list, bindLink );
         }
 
-        private void bindLink( IBlock block, int id ) {
+        private void bindLink( IBlock block, long id ) {
             block.Set( "data.LinkEdit", to( Edit, id ) );
             block.Set( "data.LinkDelete", to( Delete, id ) );
         }
@@ -37,7 +37,7 @@ namespace wojilu.Web.Controller.Download.Admin {
         [HttpPost]
         public virtual void SaveSort() {
 
-            int id = ctx.PostLong( "id" );
+            long id = ctx.PostLong( "id" );
             String cmd = ctx.Post( "cmd" );
 
             FileCategory acategory = FileCategory.GetById( id );
@@ -81,7 +81,7 @@ namespace wojilu.Web.Controller.Download.Admin {
             echoToParentPart( lang( "opok" ) );
         }
 
-        public void Edit( int id ) {
+        public void Edit( long id ) {
             target( Update, id );
 
             FileCategory cat = FileCategory.GetById( id );
@@ -94,7 +94,7 @@ namespace wojilu.Web.Controller.Download.Admin {
         }
 
         [HttpPost]
-        public void Update( int id ) {
+        public void Update( long id ) {
 
             string name = ctx.Post( "Name" );
             if (strUtil.IsNullOrEmpty( name )) {
@@ -111,7 +111,7 @@ namespace wojilu.Web.Controller.Download.Admin {
         }
 
         [HttpDelete]
-        public void Delete( int id ) {
+        public void Delete( long id ) {
             FileCategory cat = FileCategory.GetById( id );
             if (cat != null) {
                 cat.delete();
