+++ wojilu.Apps.Download/Controller/Admin/UploadController.cs
@@ -9,7 +9,7 @@ namespace wojilu.Web.Controller.Download.Admin {
 
     public class UploadController : ControllerBase {
 
-        public void PreviewPic( int id ) {
+        public void PreviewPic( long id ) {
             target( SavePreviewPic, id );
 
             FileItem f = FileItem.findById( id );
@@ -22,7 +22,7 @@ namespace wojilu.Web.Controller.Download.Admin {
             
         }
 
-        public void SavePreviewPic( int id ) {
+        public void SavePreviewPic( long id ) {
 
             if (ctx.HasUploadFiles==false) {
                 errors.Add( "请上传图片" );
